// src/app/interfaces/user-data.ts
export interface UserData {
  address: string;
  custom: string;
  date_joined: string;
  email: string;
  first_name: string;
  groups: [];
  id: number;
  is_active: boolean;
  is_staff: boolean;
  is_superuser: boolean;
  last_login: string;
  last_name: string;
  password: string;
  phone: number;
  user_permissions: [];
  username: boolean;
}